VmStatusIcon: unify icons with VM actions
Fixes: https://github.com/oVirt/ovirt-web-ui/issues/173
