+++ src/components/VmStatusIcon/style.css
@@ -1,0 +1,3 @@
+.green {
+    color: #24d011;
+}
